+++ requirements.txt
@@ -1,0 +1,3 @@
+pymongo==4.2.0
+sanic==22.3.2
+python-dotenv==0.21.0
