+++ src/GameCodin/app/app_routing.py
@@ -1,0 +1,7 @@
+from sanic import text
+
+from . import app
+
+@app.get('/users')
+async def handler(request):
+    return text("Ok")
