+++ src/GameCodin/game_room/game_room.py
@@ -1,6 +1,6 @@
 from game_room.game_room_config import GameRoomConfig
 from puzzle.puzzle import Puzzle
-from src.user.user import User
+from ..user.user import User
 from dataclasses import dataclass, asdict, field
