+++ src/GameCodin/user/user.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass, asdict, field
-from typing import Dict
 
 
 @dataclass
@@ -7,8 +6,8 @@ class User:
     username: str
     email: str
 
-    profile: Dict = field(default_factory=dict)
-    rank: Dict = field(default_factory=dict)
+    profile: dict = field(default_factory=dict)
+    rank: dict = field(default_factory=dict)
 
     @property
     def dict(self):
