+++ src/GameCodin/utils/utils.py
@@ -6,9 +6,7 @@ class Proto(Protocol):
     def static_init(cls):
         pass
 
-
 P = TypeVar("P", bound=Proto)
-
 
 def static_init(cls: P) -> P:
     cls.static_init()
